# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API",
        "description": "Descripción de la API",
        "contact": {
            "email": "[email protected]"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor API"
        },
        {
            "url": "/api",
            "description": "Servidor API"
        }
    ],
    "paths": {
        "/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener lista de productos",
                "operationId": "080f057905dc32d2f11065d606dc4cd6",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida exitosamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "eddf8c7847239ddc1040daadd814e21b",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente"
                    },
                    "500": {
                        "description": "Error al crear el producto"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener detalles de un producto",
                "operationId": "b5099368868dbe7900acd129345eb148",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del producto obtenidos exitosamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto existente",
                "operationId": "2ce632b9d1c0c445557821d09add7ec2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado con éxito"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "500": {
                        "description": "Error al actualizar el producto"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "operationId": "76bb7dc4ab3fc3233d5b33ef58e95d7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado exitosamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "500": {
                        "description": "Error al eliminar el producto"
                    }
                }
            }
        },
        "/sales": {
            "get": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Obtener lista de ventas",
                "operationId": "e9071d925184ec3b9522775e6810a95f",
                "responses": {
                    "200": {
                        "description": "Lista de ventas obtenida exitosamente"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Registrar una nueva venta",
                "description": "Almacena una nueva venta y actualiza el stock de los productos.",
                "operationId": "9741700fd7f92aca7b46ae0fec83a5e1",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Venta registrada con éxito"
                    },
                    "400": {
                        "description": "Error al registrar la venta"
                    }
                }
            }
        },
        "/sales/{id}": {
            "get": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Obtener detalles de una venta",
                "description": "Muestra los detalles de una venta específica.",
                "operationId": "4d377bdabfbfc2a1ac6f34309f497198",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la venta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la venta obtenidos exitosamente"
                    },
                    "404": {
                        "description": "Venta no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Actualizar una venta existente",
                "description": "Actualiza una venta específica.",
                "operationId": "c1f40c960f24f9afecf4f406c4d49173",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la venta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Venta actualizada con éxito"
                    },
                    "400": {
                        "description": "Error al actualizar la venta"
                    },
                    "404": {
                        "description": "Venta no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Eliminar una venta",
                "description": "Elimina una venta específica.",
                "operationId": "845ed4117f6ab26c4231dd8ff90687f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la veta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venta eliminado exitosamente"
                    },
                    "404": {
                        "description": "Venta no encontrada"
                    },
                    "500": {
                        "description": "Error al eliminar la venta"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Productos",
            "description": "Operaciones relacionadas con productos"
        },
        {
            "name": "Ventas",
            "description": "Operaciones relacionadas con ventas"
        }
    ]
}